add question reply and update login

+++ src/main/java/com/kyrie/community/controller/PublishController.java
@@ -24,7 +24,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/edit/{id}")
-    public String edit(@PathVariable(value = "id") Integer id,
+    public String edit(@PathVariable(value = "id") Long id,
                        Model model
     ) {
         QuestionDTO questionDTO = questionService.findById(id);
@@ -48,7 +48,7 @@ public class PublishController {
             @RequestParam(name = "title", required = false) String title,
             @RequestParam(name = "description", required = false) String description,
             @RequestParam(name = "tag", required = false) String tag,
-            @RequestParam(name = "id", required = false) Integer id,
+            @RequestParam(name = "id", required = false) Long id,
             Model model,
             HttpServletRequest request
     ) {
