+++ src/main/java/com/kyrie/community/dto/QuestionDTO.java
@@ -9,15 +9,15 @@ import lombok.Data;
  */
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmtCreated;
     private Long gmtModified;
-    private Integer creatorId;  // 创建者ID
-    private Integer commentCount;
-    private Integer viewCount;
-    private Integer likeCount;
+    private Long creatorId;  // 创建者ID
+    private Long commentCount;
+    private Long viewCount;
+    private Long likeCount;
     private String tag;
     private TbUser tbUser;
 }
